Add files via upload
Globally, ecosystem function is under threat due to recurring droughts and declining biodiversity, with the intricate relationship between drought and diversity-productivity remaining poorly understood. This paper aims to ensure the sustainability of plant communities in long-term, large-scale environments. The focus is on elucidating the impact of irregular weather patterns on species numbers and plant community characteristics. To achieve this, we employ a differential equation-based model to explore species interactions, offer numerical solutions, and draw relevant conclusions.

In addressing the first issue, we endeavor to forecast changes in plant communities over time, accounting for irregular weather patterns and species interactions. We begin by analyzing the various types of species interactions using niche and competitive differences, deriving coefficients for these interactions through the jungle model. We then simulate rainfall patterns, constructing a Lotka-Volterra competition model based on differential equations. The outcome is a curve depicting changes in species populations over time.

For the second problem, an optimization challenge, we seek to determine the minimum number of species required to benefit the community. To assess the community's benefit, we rely on five indicators related to plant community characteristics. Our analysis reveals that these indicators reach optimal and stable values when the number of species reaches five.

In tackling the third issue, we explore the influence of species types on the results. Species are categorized into three groups based on their interactions, and a series of experiments are conducted under controlled conditions. The findings highlight that inferior species thrive in the absence of dominant ones, while neutral species are minimally affected by dominant species.

The fourth problem involves predicting the effects of drought intensification and mitigation on plant growth, quantifying the impact of different species on total individual numbers. We employ random number generation to fit adjusted rainfall patterns and solve equations based on differential equations. The results indicate that increased drought leads to a reduction in most species, while reduced drought favors the growth of the majority. Notably, drought mitigation, with the most significant overall community impact, is attributed to species PLER and HOMU, with quantified impact values of 13.4‰ and 10.8‰, respectively.

In addressing the fifth question, we forecast species growth under varying nitrogen dioxide concentrations and environmental carrying capacity. Utilizing a logistic curve model to capture the impact of nitrogen dioxide concentration on plant growth rates, we simulate habitat changes by adjusting the environmental carrying capacity parameter. The outcome is represented by differential equations solving for plant growth.

Finally, leveraging our model, we offer recommendations for the long-term and large-scale development of plant communities while conducting an analysis of the model's strengths and weaknesses.